(test): Add more unit and integration tests
Looking at how the unit tests are built, trying to get more in-depth
with the template assertions. The added test for testing teamcity will
double-check that ports are correct in the template.

Additionally started the structure for integration testing at the end
stage of the Pipeline, run through Lambda functions that will test
stacks